Finish up for v0.3, documentation and module separation
